Add pixel-grid alignment hints
    <path id="align-anchor" d="m #,# ..." />

The first (move-to) point in the path is used to align to the target pixel
grid.  For example:

    <svg width="256" height="256" viewBox="0 0 256 256">
      <path id="align-anchor" d="m117.44 114.88 ..." />
      ...

Preparing to output a 48×48 image with view-box expanded to 384 (1.5 × 256,
Windows cursor regular size scheme):

    <svg width="48" height="48" viewBox="0 0 384 384">
      ...

the offset is calculated as follows:

    scaleX = targetSize.width / viewBox.width
    scaleY = targetSize.height / viewBox.height

    // Assuming source view-box always starts at 0,0
    alignX = (int) round(anchor.x * scaleX)
    alignY = (int) round(anchor.y * scaleY)
    alignX /= scaleX
    alignY /= scaleY

    offsetX = anchor.x - alignX
    offsetY = anchor.y - alignY

The offsets (that may be negative) are then used for the view-box origin:

    <svg width="48" height="48" viewBox="-2.56 2.88 384 384">
      ...

The "align-anchor" should get perfectly aligned to the target resolution
pixel grid.