Add anchor (incl. hotspot) biases
Add align-anchor (and cursor-hotspot) bias allowing to adjust the coordinates
in case of dynamic geometry changes such as changing shape's stroke-width
(could there be any other?).

Rework some of the pointers or some of their details to be better suited
for dynamic stroke change - try to preserve filled areas when the stroke
shrinks (not expanding the filled area).  If all shapes are like this,
they wouldn't need bias to their alignment anchors.