Longer double-size arrows
Seems like Bibata doesn't balance the sizes of different pointers very well.
This one is the most striking one to me – the size of the main pointer shape,
especially the modern one, and the double-size arrows.  These arrows should
be generally longer than the main pointer width/height, but as currently
laid out the arrows end up quite shorty.

This is the easiest (least changes) workaround I could think of.  Requires
minimum `-s 1.25` (125% canvas sizing) to not crop the rendering.