Render the animated cursors from the animated SVG versions
Now that mousegen 0.6+ supports animated SVGs with the default JSVG backend.
It is faster to render and more convenient to maintain a single SVG source.

The configuration is adjusted so the output produces the exact frame count
as currently static SVG frames are defined, and then the animation speed is
preserved:

    durationSeconds(3) * frameRate(14.333) = 43 [frames]

    1000 / (frameRate(14.333) * rateGain(1.395)) = 50 [ms]

    60 / (frameRate(14.333) * rateGain(1.395)) = 3 [jiffies]