Hotspots on the inside edge of the stroke
More precisely – just outside the fill.  This matches the setup of KDE Breeze
cursors, for example.

When the stroke and the background are the same color it effectively becomes
invisible.  Then the outer edge of the fill is the visual guide.  In this
situation it could be confusing if the cursor "activates" a bit before it reaches
the intended target.  This would be more apparent with thicker pointer outlines.

When the stroke and the background are in contrast, it may appear as if the pointer
"activates" a tiny bit later.  I still think this is the less confusing behavior.
The hotspot of many pointers is inside the shape (such a with the "crosshair")
so it is likely users anticipate this "better".